refactor: move common function to ipfs utils
