+++ components/collection/CollectionList.vue
@@ -55,7 +55,7 @@ import {
   Metadata,
   NFTMetadata,
 } from '@/components/rmrk/service/scheme'
-import { getSanitizer } from '@/components/rmrk/utils'
+import { getSanitizer } from '@/utils/ipfs'
 import { SearchQuery } from '@/components/rmrk/Gallery/search/types'
 import 'lazysizes'
 import collectionListWithSearch from '@/queries/subsquid/general/collectionListWithSearch.graphql'
