+++ components/gallery/useGalleryItem.ts
@@ -1,5 +1,5 @@
 import { $fetch } from 'ohmyfetch'
-import { sanitizeIpfsUrl } from '@/components/rmrk/utils'
+import { sanitizeIpfsUrl } from '@/utils/ipfs'
 import { getMimeType } from '@/utils/gallery/media'
 import type { NFT, NFTMetadata } from '@/components/rmrk/service/scheme'
