+++ pages/westmint/u/_id.vue
@@ -131,7 +131,7 @@
 import { Component, Watch, mixins } from 'nuxt-property-decorator'
 
 import { CollectionWithMeta, Pack } from '@/components/rmrk/service/scheme'
-import { fetchNFTMetadata, sanitizeIpfsUrl } from '@/components/rmrk/utils'
+import { fetchNFTMetadata, sanitizeIpfsUrl } from '@/utils/ipfs'
 
 import { notificationTypes, showNotification } from '@/utils/notification'
 import PrefixMixin from '@/utils/mixins/prefixMixin'
